# [MISC] Use non-blocking transfer in prepare_input