This PR uses non-blocking data transfer in prepare_input. This is beneficial because we transfer several tensors to GPU in prepare_input. Here are some benchmark results using Llama-3.1-8B-Instruct on 1xH100:

Batching

Command:
```
python3 benchmark_throughput.py \
    --model meta-llama/Meta-Llama-3.1-8B-Instruct \
    --backend vllm \
    --input-len 292 \
    --output-len 579 \
    --num-prompts 1000
```
Result (I observed some variants so if you ran this multiple times the throughput is actually ranging from 8.15~8.34).
```
Main:    Throughput: 8.12 requests/s, 7075.74 tokens/s
This PR: Throughput: 8.34 requests/s, 7266.01 tokens/s
```

Serving

I used a different benchmark framework so no commands here, but the settings are as follows:
- Input / Output: 550 / 150
- QPS: 8
- Duration: 120 seconds
```
Main:    P90 Latency: TTFT 47.9ms, ITL 12.9ms, E2E 2.0s
This PR: P90 Latency: TTFT 44.8ms, ITL 11.5ms, E2E 1.8s
```

Reference: https://pytorch.org/tutorials/intermediate/pinmem_nonblock.html

cc @youkaichao